Context and interface now two separate files
